+++ src/ts_catalog/catalog.h
@@ -58,6 +58,7 @@ typedef enum CatalogTable
 	CONTINUOUS_AGGS_BUCKET_FUNCTION,
 	JOB_ERRORS,
 	CONTINUOUS_AGGS_WATERMARK,
+	TELEMETRY_EVENT,
 	/* Don't forget updating catalog.c when adding new tables! */
 	_MAX_CATALOG_TABLES,
 } CatalogTable;
@@ -885,6 +886,22 @@ enum
 	METADATA_PKEY_IDX = 0,
 	_MAX_METADATA_INDEX,
 };
+
+/*
+ * telemetry_event table definition
+ */
+
+#define TELEMETRY_EVENT_TABLE_NAME "telemetry_event"
+
+enum Anum_telemetry_event
+{
+	Anum_telemetry_event_created = 1,
+	Anum_telemetry_event_tag,
+	Anum_telemetry_event_body,
+	_Anum_telemetry_event_max,
+};
+
+#define Natts_telemetry_event_max (_Anum_telemetry_event_max - 1)
 
 /****** BGW_POLICY_CHUNK_STATS TABLE definitions */
 #define BGW_POLICY_CHUNK_STATS_TABLE_NAME "bgw_policy_chunk_stats"